encapsulate validation functions
